python_ulid_benchmark/pyproject.toml
[project]
name = "python-ulid-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "psutil>=7.1.0",
    "python-ulid>=3.1.0",
    "seaborn>=0.13.2",
    "typing-extensions>=4.15.0",
    "ulid-python>=1.0.1",
]
